Removed deprecated APIs
This change removes the APIs that were deprecated in v0.10.0-preview and
outlines the plan for the rest of 0.11.0-preview.
