+++ packages/network-console-shared/changelog.md
@@ -5,6 +5,28 @@ Network Console hosts and clients. Because this is a shared component,
 version changes might be breaking. This changelog will attempt to
 document when breaking changes occur.
 
+## v0.11.0-preview
+
+This change removes the deprecated Collections and Environments APIs:
+
+ - `Collections.Postman21NativeReader`
+ - `Collections.NCNativeReader`
+ - `Collections.SwaggerFileFormatReader`
+ - `Collections.tryReadCollectionAsync`
+ - `Environments.tryReadEnvironmentAsync`
+ - `Environments.serializeNativeEnvironmentFormat` 
+ - `CollectionItemType`, `ICollectionEntryReader`, `ICollectionFolderReader`, 
+   `ICollectionItemBase`, `ICollectionRootReader`, `ICollectionsReader`
+ - `EnvironmentItemType`, `IEnvironment`, `IEnvironmentRoot`
+
+Additionally, this revision introduces the i18n module, powered by the
+same component that Google's Lighthouse uses:
+https://github.com/GoogleChrome/lighthouse/tree/master/lighthouse-core/lib/i18n
+
+Finally, this change introduces a refactor that produces a new, updated
+ES Modules export shape because of changes to Chromium and Edge for how
+the ESM engineering systems changes have occurred.
+
 ## v0.10.0-preview
 
 This change introduces a new File I/O module. This module is intended
