+++ packages/network-console-shared/src/file_io/native/collection-format.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 import { ICollectionFormat, ICollectionAdapter } from '../interfaces';
-import { INCNativeRoot } from '../../collections/native/native-file-format';
+import { INCNativeRoot } from './format';
 import { CollectionAdapter } from './collection-adapter';
 
 export class CollectionFormat implements ICollectionFormat {
