+++ packages/network-console-shared/src/util/loc.ts
@@ -1,0 +1,3 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
